Allow users to make requests (#27)
To abstract away from the notifier script and to be able
to use similar functions in the monitor requests script
Notify was pulled out to do all the common emailing functions

Added a python monitor requests script that has an on snapshot of the requests collection on firebase. An email is sent and the entry is deleted immediately after being created. Similarly, the website posts to this same collection which then creates a request entry for the script.